X-NWB DatConverter/ABFConverter: Introduce cycle_id property
When reading the NWB files it is currently not possible to group
TimeSeries back together to the same sweep from which they originated
from.

See https://github.com/NeurodataWithoutBorders/pynwb/issues/499 for a
proposal to solve that within pynwb.

For now we work around that by introducing a "cycle_id" property which
is contstant for TimeSeries which belong together.