DatConverter: Handle different AmplifierState locations
The release notes of v2x90.3, 19-Mar-2018 has the following entry[1]:

The amplifier settings are now always stored in a separate "*.amp" file
(which is part of the DAT bundle file). In previous versions, such a
separate file was created only for EPC/n amplifiers.

[1]: http://www.heka.com/downloads/software/release_notes/release_notes.html#rn_pm_v2x9

and therefore we must also look into that location. Instead of
making the decision dependent on the version we check if the settings
are valid instead.