X-NWB/nwb_to_pdf.py: Add mode to visualize the stimset reconstruction accuracy
Both for ABF and DAT files we reconstruct the stimulus sets from their
vendor specific parameters. As that is mostly reverse engineered we need
a way to verify what we've done.

The script nwb_to_pdf.py learned to accept --check-stimset-rec which
creates three plots per page for each cycle ID. On each page where is a
plot of acquired data, the absolute difference and the relative
difference.

This mode requires NWB data with two electrodes/headstages where the
second is just a loopback.