X-NWB/mcc_get_settings.py: Add amplifier uids and make more configurable
User experience has shown that the name of the created file needs to be
tweakable. This is now possible.

In addition we need to make a connection between the name of the
amplifier "x00830251_1" and the ADC channel "IN 1". This information can
be passed in via `--idChannelMapping`.

uids idea courtesy of Djai Heyer.