NwbReader: Fix usage for v2 files
In 0a16fcf0 (nwb_reader: Prepare for classes with NWBv2 support,
2018-11-26) the NwbReader was teached to handle NWBv2 files as well. One
part of that was to abstract away the acquisition and stimulus paths.
Unfortunately that commit was not correct as it broke v1 files as the
stimulus path is not the same for both MIES and Pipeline NWBv1 flavours.
A latter commit  3335c572 (Repair nwb reader, 2018-11-30) tried to fix
that but broke v2 files again.

Now we define the acquisition_path and stimulus_path in each child
constructor and also define the NWB major version there.