NwbReader: Check that the assumed and the real sweep number are the same
Since forever the numeric value in the TimeSeries group was used as the
sweep number. For example from the group name "ABCD_123" the sweep
number 123 was derived.

For data acquired with MIES this assumption is definitly false if one of
the following conditions is met:
- TTL channels are used
- More than one active headstage
- Unassociated AD/DA channels (aka channels not connected with a headstage)

Since forever the sweep number is encoded in the source
attribute by IPNWB [1].

Introduce the function check_sweep_number which checks that the assumed
and real sweep number match. This function is called for both MIES and
Pipeline NWB files in the get_sweep_data routine.

[1]: https://github.com/AllenInstitute/IPNWB#nwb-file-format-description