Tests EphysDataSet: Add tests for get_sweep_info_by_sweep_number and adapt default data
The default data returned in the dictionary by get_sweep_table_dict()
has None for invalid numeric entries.

As the test shows this is implicitly converted to NaN somewhere in
the pandas Dataframe->python->pandas DataFrame conversion cycle.

Initialize these entries to np.nan to make it explicit and the test
pass.