NwbReader.get_real_sweep_number: Return the real sweep number
In 44ea93e7 (NwbReader: Check that the assumed and the real sweep number
are the same, 2018-12-06) we added code to check that the assumed sweep
number is the same as the real sweep number.

But we can be better than that and just return the real sweep number if
requested. For compatiblity with files without real sweep number info we
fallback to the assumed sweep number if possible.

See 44ea93e7 (NwbReader: Check that the assumed and the real sweep
number are the same, 2018-12-06) for further information on the real
vs assumed sweep number.