X-NWB/DatConverter: Use correct clamp mode
Due to misinterpretation of the author of this commit we used the wrong
heuristics for determining the clamp mode.

This has now been fixed and we use the Mode entry of the AmplifierState.
As we not always have an amplifierState entry we then fallback, with a
warning, to just use the AD units.

This also prevents the bug that we have different clamp modes for the
DA and the AD channel.