X-NWB/DAT/Segments: Don't fail on logarithmic incrementation modes
Due to the definition of these modes we also had to change the way
_applyDelta works, now one needs to pass in the number of sweeps instead
of calling it multiple times.

This also fixes a bug where the factor was applied too often for the
"Inc" mode.

As it is currently not clear how V * factor and dV * factor can be
distinguished we return NaN for now.