Change behavior when no current clamps are found
Stop raising exception when no sweeps are found to compute QC features.
Instead, log the warning and add a fail_tag. Subsequently, tag this cell as
failing qc.
This prevents module failure when sweeps for QC are not available, but
also allows lims to fail the cell for subsequent analysis.