mcc_get_settings.py: Add option to watch a folder
PClamp does not allow to call the mcc_get_settings.py script each time a
new ABF file is created. But we need to call that at this point in time
to fetch the current settings.

Add a --watchFolder mode which gathers the settings if a new ABF file is
created.

Uses the third party watchdog package from https://github.com/gorakhargosh/watchdog
under Apache License 2.0.