Build sweep mapping between sweep_name and sweep_number
To read sweep_data we must provide the name of the dataset
Previously it was inferred based on the assumed_sweep_number. However,
when sweeps are reacquired for whichever reason, the assumed_sweep
number is not longer equal the sweep_number and cannot be used. Thus, it
became necessary to build a mapping between the real sweep_number and the
sweep_name. Notably the sweep_name is the name of the acquisition data
set.

Move dropping earlier recorded sweep with the same sweep number into
nwb_reader.