GH #485 - Fixing truncated square pulse error
If the sequence of square pulses starts 'downward' or ends 'upward',
the get_square_pulse_idx function raises a 'Truncated square pulse'
error. However, we could use the remaining pulses to complete the
calculation. We introduced a small modification that allows dropping
the surplus downward or upward parts of truncated square pulses and
move forward. If we still have an unequal number of up and down
changes, there must be another issue with the pulses, so an error is
raised with the message 'Invalid square pulse.'