Fix missing electrode 0 when in bath sweep was ran with smoketest stimulus
This commit extends the 'extract_electrode_0' function and tries to
determine if the 'in-bath' sweep was ran with the 'EXTPSMOKET' stimulus
when the 'EXTPINBATH' stimulus is not found.

If 'EXTPINBATH' is not found then measure the seal of all 'EXTPSMOKET'
sweeps that occur before the first instance of 'EXTPCllATT' and only if
the resistance of the seal is below 10 MOhms then call
'measure_electrode_0'. Electrode 0 is not available if no 'EXTPSMOKET'
sweeps occurred before the first 'EXTPCllATT' sweep, the resistance is
above 10 MOhms, or if no 'EXTPSMOKET' or no 'EXTPCllATT' sweeps found.