Refactor extract_electrode_0 and new functions for improved testing
This commit refactors the code that was added to extract_electrode_0 for
finding inbath sweeps ran with the smoketest stimulus. The previous unit
tests were not hermetic so this refactor and the couple of new functions
allow for hermetic testing.

new functions: filter_smoketests filters out any smoketest sweeps that
occurred after the cell attached sweep and minimum_e0 finds the minimum
absolute value from a list of e0 values; these e0 values are measured
from the smoketest sweeps that occurred after the first cell attached
sweep if the resistance of the pipette is less than 10 MOhms.