GH #508 - Remove np.nan values from np.diff array
np.nan values were not being dropped from the end of the np.diff array
before calling np.flatnonzero to remove non-zero values in get_stim_epoch.
This was causing get_stim_epoch to incorrectly identify the end of the
stimulus epoch as the same as the end of the sweep epoch, which was also
causing the end of the experiment epoch to be incorrect as well since it
relys on the stimulus epoch. By dropping the np.nan values from the end
of the difference array (di) in get_stim_epoch, the correct indices for
end of the stimulus epoch as well as for the experiment epoch are now being
returned. This should fix the issue with sweeps being incorrectly tagged
with 'Recording stopped before completing the experiment epoch'.