Only check stimulus epoch for 'Stimulus contains NaN values' tag
The check_sweep_integrity function was checking the entire sweep for NaN
values and applying the tag 'Stimulus contains NaN values; if any are found.
This has been causing most of the sweeps to fail out before sweep qc, and
when every sweep in an experiment contains this tag then the entire cell
automatically fails qc with the fail tag 'No sweep states available'. This
commit narrows the search for NaN values to only the 'stimulus' epoch of
the sweep. Almost every sweep in most Ephys experiments contain some amount
of NaN values at the end because we are auto-terminating sweeps once the
post-stimulus baseline hsa returned to within +/- 1 mV of initial baseline.