Check if stimulus end index is less than stimulus start index
If the recording is stopped before completing the stimulus epoch then
the ending index of the stimulus epoch will be less than the starting
index of the stimulus epoch and with the changes from my last commit,
this would prevent the sweep from being tagged with 'Stimulus contains
NaN values' because there would be no stimulus epoch to check for NaN's.
This commit checks the indices of the start and end of the stimulus epoch
and if the end index is less than the start index then the stimulus epoch
was not completed and the sweep is tagged.