max_pipette_resistance is now a constant in units of MOhms
This commit address the issue of the 0.01 threshold for pipette resistance.
This threshold has been moved to a constant called max_pipette_resistance
and is now in units of MOhms and is set to 7.5. Actual pipette resistance
is now found with measure_input_resistance from qc_features.py, which
returns the resistance in units of MOhms.