GH #537 - Decode nwb_version if byte object
Add get_string_value() function to check if nwb_version attribute is
an instance of byte class and decode to string if necessary. This will
prevent a TypeError crash when get_nwb_version() calls re.match() with
nwb_version and is expecting a string and not a utf-8 encoded byte object.