Fix Nwb2Sink.serialize errors / remove set_container_sources()
After updating PyNWB, the pre-existing serialize method in the Nwb2Sink
class was throwing errors along the form of:

ValueError: Cannot change container_source once set: 'root' pynwb.file.NWBFile

This commit resolves this error by removing the use of set_container_sources()

After removal of the set_container_sources() method, tests pass and
testing done by scientists also did not reveal any issues, but because
the origin/reason for this code is not clear, we're not sure that
removing it doesn't break something else with NWB files, for example
with publishing to DANDI.

Should you run into issues with containers or other metadata in NWB
files, the work of the removed code might be the explanation. However
it's not clear how to modify that code to get around the original error
mentioned above.