treewide: Allow to inject a SweepTable
The SweepTable was the first icephys table for grouping different patch
clamp series together, see NeurodataWithoutBorders/pynwb@21f91920 (Add a
SweepTable for icephys data, 2018-11-02) for pynwb 0.6.0.

This was superseeded by the icephys metadata tables introduced in pynwb
2.0.0 and finally removed in 3.0.0.

While migrating to the icephys metadata tables is on the list [1], we
don't want to do this here now.

So let's add a nasty hack for injecting a sweep table.

[1]: https://github.com/AllenInstitute/IPNWB/issues/47