Move to pyproject.toml packaging
The new way of packaging in python is based on pyproject.toml. This allows
us to have only one file defining the package properties instead of three.

We also switch to hatch instead of setuptools. And the version number is
now gathered from git instead of being hardcoded in ipfx/version.txt.

The new source package (sdist) has the following differences:
- We don't ship the tests folder anymore. It is not usable without
  datafiles anyway and including them would require ~250MB of disc space.
- `ipfx/version.txt` is gone as the version is now gathered from git with the build-time created `ipfx/_version.py`
- No more IPFX.egg-info/MANIFEST.in/setup.cfg/setup.py but only pyproject.toml
- Added missing files in docs

The wheel is now different in the following regard:
- It is not an universal wheel anymore as Python 2 support is already long gone
- No more `*dist-info` folder as we are now packaging with hatch, the files LICENSE and AUTHORS.rst are
  therefore manually included
- Same comment regarding `ipfx/version.txt` as in the sdist

The flake8 configuration can not be included in pyproject.toml, see
https://github.com/PyCQA/flake8/issues/234. It will also be
replaced by another linter in a future commit.