test_serialize: Silence pynwb warning
pynwb 3.1.2 does currently always output a warning about a missing .nwb
ending in path even if path was not passed at all and is None.

This will be fixed in [1], ignore the warning until we require that
version.

We also drop the path argument to NWBHDF5IO as that is not needed.

[1]: https://github.com/NeurodataWithoutBorders/pynwb/pull/2130