treewide: Get on-premise tests working again
The on-premise tests using the python 3.9 and 3.11 tests are currently not
working due to permission issues around the CWD when executing docker run.

The reason is that the user id of the user inside the docker image and of
that on the machine running github actions is different.

By now creating a user with uid/gid 1001 we match what is on the host.

But the images are currently only rebuild on demand manually on the
machine. We fix that by building them before the run step.

The run step is also moved to a script for easier local testing and to
avoid code duplication.

And finally the Dockerfile is used with a ARG'ed FROM to avoid
duplication.