ipfx/lims_queries.py: Fix racy dictionary querying
The on-premise tests were fixed in 078bda8 (Merge pull request #581 from
AllenInstitute/fix-on-premise-tests, 2025-08-26) and also passed on
master [1]. Now they are not passing again with an error like:

credentials = dict((k, os.environ.get(env_var, LIMS_DB_CREDENTIAL_DEFAULTS[env_var]))
        for k, env_var in LIMS_DB_CREDENTIAL_MAP.items())
KeyError: 'LIMS_PASSWORD'

The dict LIMS_DB_CREDENTIAL_DEFAULTS does not contain an entry with
"LIMS_PASSWORD" as key which should be okay as we are always having that
in the envrionment. But this seems to be racy as sometimes we try indexing
LIMS_DB_CREDENTIAL_DEFAULTS with "LIMS_PASSWORD" anyway.

Let's fix that by adding an entry do LIMS_DB_CREDENTIAL_DEFAULTS.

[1]: https://github.com/AllenInstitute/ipfx/actions/runs/17247644153